refactored the createpixels function to be in a different file
